Move sacrifice precompile to its own file
